% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptools2-package.R
\docType{package}
\name{iptools2}
\alias{iptools2}
\alias{iptools2-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/iptools2}
\item BugReports: \url{https://gitlab.com/hrbrmstr/iptools2/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
