% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{range_boundaries}
\alias{range_boundaries}
\title{Retrieve range boundaries for a character vector of IPv4/IPv6 CIDRs}
\usage{
range_boundaries(cidr)
}
\arguments{
\item{cidr}{a character vector of IPv4/IPv6 CIDRs}
}
\value{
data frame with original input, canonical CIDR, start/end (chr) of range and number of hosts.
}
\description{
Retrieve range boundaries for a character vector of IPv4/IPv6 CIDRs
}
\note{
the \code{n_hosts} column is a numeric vector since R cannot handle 128-bit integers easily and
v4 & v6 are co-mingled.
}
\examples{
range_boundaries(
  c("2001:4801::/32", "2001:0db8::/127", "10.1.10.0/24", "2001:0db8::/119", "wat")
)
}
