% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptools2-package.R
\docType{package}
\name{iptools2}
\alias{iptools2}
\title{A WIP Re-imagining of {iptools}}
\description{
A WIP Re-imagining of {iptools}
}
\details{
\itemize{
\item URL: \url{https://git.rud.is/hrbrmstr/iptools2}
\item BugReports: \url{https://git.rud.is/hrbrmstr/iptools2/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
